Renamed "api" to "all" for consistency.
