+++ open_mafia_engine/builders/for_testing.py
@@ -1,5 +1,5 @@
 from typing import List
-from open_mafia_engine.core.api import *
+from open_mafia_engine.core.all import *
 from open_mafia_engine.built_in.all import *
