+++ open_mafia_engine/built_in/voting.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 import logging
 from typing import DefaultDict, Dict, List, Optional, Tuple, Union
 
-from open_mafia_engine.core.api import (
+from open_mafia_engine.core.all import (
     Ability,
     Action,
     Actor,
